moved emptyS3Bucket/DeleteS3Bucket from cloudformation.go to purge.go.
Also introduced DeleteAnyS3Buckets and DeleteAnyEcrRepos to move the for-loops that used to surround the emptyS3Bucket/DeleteS3Bucket functions, so they are also moved to purge.go.
